convert events to classes
