+++ src/app/events/Ready.ts
@@ -1,16 +1,12 @@
-/**
- * Handler for Discord event `ready`.
- */
-
 import GuestPassService from '../service/guest-pass/GuestPassService';
 import { Client } from 'discord.js';
 import constants from '../service/constants/constants';
 import { connect } from '../utils/db';
+import { Event } from '../types/Event';
 
-
-module.exports = {
-	name: 'ready',
-	once: true,
+export default class implements Event {
+	name = 'ready';
+	once = true;
 
 	async execute(client: Client) {
 		console.log('The Sun will never set on the DAO. Neither will I. DEGEN & Serendipity are ready for service.');
@@ -18,5 +14,5 @@ module.exports = {
 		await connect(constants.DB_NAME_DEGEN);
 		await connect(constants.DB_NAME_BOUNTY_BOARD);
 		await GuestPassService(client);
-	},
+	};
 };
