+++ src/app/events/welcomeMats/GuildMats.ts
@@ -1,4 +1,4 @@
-import devGuildMat from './devGuild';
+import devGuildMat from './DevGuild';
 
 export default Object.freeze({
 	devGuildMat: devGuildMat,
