+++ src/app/types/Event.ts
@@ -1,0 +1,5 @@
+export interface Event {
+    name: string,
+    once: boolean,
+    execute(...args: any[]): void
+}
